multi-agent-manufacturing/requirements.txt
crewai
crewai-tools
langchain-groq
litellm
python-dotenv
ddgs
langchain-community
python-docx
streamlit